windsurf-project/requirements.txt
# Hidden Job Market Intelligence System
# Python 3.9+

# HTTP & Scraping
requests>=2.31.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selectolax>=0.3.0

# Rate Limiting & Async
aiohttp>=3.9.0
tenacity>=8.2.0
ratelimit>=2.2.1

# Data Processing
pandas>=2.1.0
pyyaml>=6.0.1

# Database
sqlite-utils>=3.35.0

# URL & Domain Handling
tldextract>=5.1.0
urllib3>=2.1.0

# Text Processing
rapidfuzz>=3.5.0

# CLI & Output
rich>=13.7.0
click>=8.1.0

# Optional: Change Detection
schedule>=1.2.0
